+++ test-packages/typed-json-model/package.json
@@ -16,7 +16,7 @@
     "ci": "npm run lint && npm run ui5lint && npm run ts-typecheck && npm run test"
   },
   "devDependencies": {
-    "@types/openui5": "1.136.0",
+    "@openui5/types": "^1.146.0",
     "@ui5/cli": "^4.0.30",
     "@ui5/linter": "^1.20.2",
     "eslint": "^9.37.0",
